requirements.txt
Django>=4.2
djangorestframework
celery>=5.3.0
redis>=5.0.0
mysqlclient

sentence-transformers
qdrant-client
openai

minio
PyMuPDF==1.23.24

faker
django-cors-headers
